refactor: extract activity timer modules
